backend/requirements.txt
strands-agents
strands-agents-tools
ag-ui-strands
neo4j
fastapi
uvicorn
python-dotenv
python-multipart
requests
PyPDF2
edge-tts
openai
ollama
tavily-python
# Optional telemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
langfuse